[evo] #1600 - gérer la documentation à destination des déposants et élus
 - New component for admin documentations
 - New component for Patchs
 - New component for Helps
 - New api endpoint for documentations
 - New type form for patchs and helps
 - Upgrade translations for admin documentations

+++ src/Sesile/MainBundle/Entity/Aide.php
@@ -52,7 +52,7 @@ class Aide
     /**
      * @var \DateTime
      *
-     * @ORM\Column(name="date", type="datetime")
+     * @ORM\Column(name="date", type="datetime", options={"default": "CURRENT_TIMESTAMP"}))
      *
      *
      */
@@ -153,7 +153,7 @@ class Aide
     public function removeUpload()
     {
         if ($file = $this->getAbsolutePath()) {
-            unlink($file);
+            if(is_file($file))  unlink($file);
         }
     }
